+++ lib/PhpParser/Builder/Namespace_.php
@@ -6,7 +6,7 @@ use PhpParser;
 use PhpParser\Node;
 use PhpParser\Node\Stmt;
 
-class Namespace_ extends PhpParser\BuilderAbstract
+class Namespace_ extends Declaration
 {
     private $name;
     private $stmts = array();
@@ -34,26 +34,11 @@ class Namespace_ extends PhpParser\BuilderAbstract
     }
 
     /**
-     * Adds multiple statements.
-     *
-     * @param array $stmts The statements to add
-     *
-     * @return $this The builder instance (for fluid interface)
-     */
-    public function addStmts(array $stmts) {
-        foreach ($stmts as $stmt) {
-            $this->addStmt($stmt);
-        }
-
-        return $this;
-    }
-
-    /**
      * Returns the built node.
      *
      * @return Node The built node
      */
     public function getNode() {
-        return new Stmt\Namespace_($this->name, $this->stmts);
+        return new Stmt\Namespace_($this->name, $this->stmts, $this->attributes);
     }
 }