+++ test/PhpParser/Builder/NamespaceTest.php
@@ -2,6 +2,7 @@
 
 namespace PhpParser\Builder;
 
+use PhpParser\Comment\Doc;
 use PhpParser\Node;
 use PhpParser\Node\Stmt;
 
@@ -15,19 +16,23 @@ class NamespaceTest extends \PHPUnit_Framework_TestCase
         $stmt1 = new Stmt\Class_('SomeClass');
         $stmt2 = new Stmt\Interface_('SomeInterface');
         $stmt3 = new Stmt\Function_('someFunction');
+        $docComment = new Doc('/** Test */');
         $expected = new Stmt\Namespace_(
             new Node\Name('Name\Space'),
-            array($stmt1, $stmt2, $stmt3)
+            array($stmt1, $stmt2, $stmt3),
+            array('comments' => array($docComment))
         );
 
         $node = $this->createNamespaceBuilder('Name\Space')
             ->addStmt($stmt1)
             ->addStmts(array($stmt2, $stmt3))
+            ->setDocComment($docComment)
             ->getNode()
         ;
         $this->assertEquals($expected, $node);
 
         $node = $this->createNamespaceBuilder(new Node\Name(array('Name', 'Space')))
+            ->setDocComment($docComment)
             ->addStmts(array($stmt1, $stmt2))
             ->addStmt($stmt3)
             ->getNode()