+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 3.1.2-dev
 
 * Added `kind` attribute for `Stmt\Namespace_` node, which is one of `KIND_SEMICOLON` or
   `KIND_BRACED`. (#417)
+* Added `setDocComment()` method to namespace builder. (#437)
 
 Version 3.1.1 (2017-09-02)
 --------------------------